requirements.txt
requests==2.31.0
python-telegram-bot==20.6
python-dotenv==1.0.0
dataclasses-json==0.6.1
schedule==1.2.0
aiohttp==3.9.1
typing-extensions==4.8.0
pydantic==2.5.0
flask==3.0.0 